feat: create custom hook for auth validation
