+++ src/views/Authentication/Authentication.tsx
@@ -5,16 +5,25 @@ import {
   useLoginMutation,
   useRegisterMutation,
 } from '../../redux/user/userApiSlice';
-import { validEmail, validPassword } from './authHelper';
+import { useValidateForm } from './useValidateForm';
 import { useFormErrorInfo } from './useFormErrorInfo';
 
 type AuthProps = {
   isRegisterForm: boolean;
 };
 
+type FormData = {
+  name: string;
+  username: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+};
+
 const Authentication = (props: AuthProps) => {
   const [login, loginData] = useLoginMutation();
   const [register, registerData] = useRegisterMutation();
+  const { formError, invalidForm, validateFormData } = useValidateForm();
 
   const loginStatus = { isError: loginData.isError, error: loginData.error };
   const registerStatus = {
@@ -22,16 +31,12 @@ const Authentication = (props: AuthProps) => {
     error: registerData.error,
   };
 
-  const [formData, setFormData] = useState({
+  const [formData, setFormData] = useState<FormData>({
     name: '',
     username: '',
     email: '',
     password: '',
     confirmPassword: '',
-  });
-  const [formError, setFormError] = useState({
-    error: false,
-    message: '',
   });
 
   const handleFormChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -41,45 +46,10 @@ const Authentication = (props: AuthProps) => {
     }));
   };
 
-  const invalidForm = (
-    message: string = 'Please make sure to fill out form properly.',
-  ) => {
-    setFormError({ error: true, message });
-  };
-
-  const validateFormData = () => {
-    const { name, email, password, confirmPassword } = formData;
-    //validate name when registering
-    if (!name) {
-      invalidForm('Please type in your name.');
-      return false;
-    }
-    //validate email
-    else if (!email || !validEmail(email)) {
-      invalidForm('Please include a valid email address.');
-      return false;
-    } else if (!password) {
-      invalidForm('Please type in your password.');
-      return false;
-    }
-    //validate passwords match when registering
-    else if (password !== confirmPassword) {
-      invalidForm('Passwords need to match.');
-      return false;
-    } else if (!validPassword(password)) {
-      invalidForm(
-        'Password should have at least 8 characters, contain one uppercase, and one number.',
-      );
-      return false;
-    } else {
-      return true;
-    }
-  };
-
   const handleFormSubmit = () => {
     const { name, username, email, password } = formData;
     if (props.isRegisterForm) {
-      const formIsValid = validateFormData();
+      const formIsValid = validateFormData(formData);
       formIsValid && register({ name, username, email, password });
     } else {
       if (email && password) login({ email, password });
